fix: update staking version and improve documentation
The version of staking has been updated to 1.2.0 from 1.0.0. Furthermore, corrections were made to the acceptSplit documentation to fix a typographical error. Additional documentation was also added to the recoverAccount function, providing a detailed explanation on how it can be used to recover a user's stake account ownership when mistakenly created using their token account address.