Replace error with RecoverWithStake in lib.rs
The SplitWithStake error message for accounts with staked tokens has been replaced with RecoverWithStake. This change is reflected in the lib.rs file under the recover account function. This implementation guards against the possibility of double voting that could occur if an account with staked tokens were allowed to be transferred.