feat: add script to build verifiable staking program
A new script called build_verifiable_staking_program.sh has been added, along with a Dockerfile to build an image for the staking program. The script builds the image and the staking program using this image, placing the artifacts at a specified location and printing the sha256sum of the staking program.