feat: add verifiable staking build script (#427)
* feat: add script to build verifiable staking program

A new script called build_verifiable_staking_program.sh has been added, along with a Dockerfile to build an image for the staking program. The script builds the image and the staking program using this image, placing the artifacts at a specified location and printing the sha256sum of the staking program.

* fix: add /artifacts to .gitignore in staking directory

The /artifacts line has been added to the .gitignore file within the staking directory. This change ensures that generated artifacts from the newly added build script and Dockerfile are not tracked by git.

* fix: use the correct version of anchor and solana for verifiable build

* chore: update docker file docs