feat: optimize delegation state (#463)
* feat: add slash contexts

* wip

* feat: complete create slash event instruction

* fix: variable rename

* fix: error

* fix: function rename

* feat: optimize del state

* refactor: update instruction index type to u64

Update the instruction index type from u8 to u64 in the `CreateSlashEvent` struct and related functions. This change allows for a larger range of indices to be used for slash events.

Refactor the code to use `index.to_be_bytes()` when generating the seed for the slash event address in the `get_slash_event_address` function.

This commit addresses the need to support a larger number of slash events and ensures consistency in the codebase.

* refactor: update SlashEvent length calculation

Update the length calculation of the SlashEvent struct to account for the new field sizes. This ensures that the length is correctly calculated and matches the expected value.

* feat: check slash custody is pyth token account