feat: add initial slash instruction (#466)
* feat: add initial slash instruction

* fix: slashing logic

* feat: improve slash test

* feat: optimize delegation state

Refactor the code to optimize the delegation state by adding slash contexts and updating the instruction index type to u64. This allows for a larger range of indices to be used for slash events. Additionally, the length calculation of the SlashEvent struct has been updated to account for the new field sizes. This commit addresses the need to support a larger number of slash events and ensures consistency in the codebase. Also, check that the slash custody is a Pyth token account.

Fixes #463

* fix: pr comments